--- a/Gemfile
+++ b/Gemfile
@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 gemspec :name => 'gollum-lib'
 gem 'irb'
-gem 'gollum-rugged_adapter', git: 'https://github.com/dometto/rugged_adapter', branch: 'improve_tree_blobs'
+gem 'gollum-rugged_adapter', git: 'https://github.com/gollum/rugged_adapter'
 
 if RUBY_PLATFORM == 'java' then
   group :development do