--- a/Gemfile
+++ b/Gemfile
@@ -3,7 +3,7 @@ gemspec :name => 'gollum-lib'
 gem 'irb'
 
 if RUBY_PLATFORM == 'java' then
-  gem 'gollum-rugged_adapter', git: 'gollum-lib_rjgit_adapter', branch: 'tree_find_blob'
+  gem 'gollum-rjgit_adapter', git: 'https://github.com/dometto/gollum-lib_rjgit_adapter', branch: 'tree_find_blob'
   group :development do
     gem 'activesupport', '~> 6.0'
   end