--- a/gollum-lib_java.gemspec
+++ b/gollum-lib_java.gemspec
@@ -1,4 +1,4 @@
 require File.join(File.dirname(__FILE__), 'gemspec.rb')
 require File.join(File.dirname(__FILE__), 'lib', 'gollum-lib', 'version.rb')
-default_adapter = ['gollum-rjgit_adapter', '~> 1.0']
+default_adapter = ['gollum-rjgit_adapter', '~> 2.0']
 Gem::Specification.new &specification(Gollum::Lib::VERSION, default_adapter, "java")