--- a/lib/gollum-lib/page.rb
+++ b/lib/gollum-lib/page.rb
@@ -10,18 +10,22 @@ module Gollum
       #
       # query     - The String path to match.
       # entry     - The BlobEntry to check against.
-      # global_match - If true, find a File matching path's filename, but not its directory (so anywhere in the repo)
-      def path_match(query, entry, global_match = false, hyphened_tags = false, case_insensitive = false)
+      def global_path_match(query, entry, hyphened_tags = false, case_insensitive = false)
         return false if "#{entry.name}".empty?
         return false unless valid_extension?(entry.name)
-        entry_name = valid_extension?(query) ? entry.name : strip_filename(entry.name)     
         match_path = ::File.join([
-          '/',
-          global_match ? nil : entry.dir,
-        entry_name
+          '/', 
+          entry.dir,
+        entry.name
         ].compact)
         path_compare(query, match_path, hyphened_tags, case_insensitive)
       end
+
+      def path_compare(query, match_path, hyphened_tags, case_insensitive)
+        cmp = valid_extension?(query) ? match_path : strip_filename(match_path)
+        super(query, cmp, hyphened_tags, case_insensitive)
+      end
+
     end
 
     # Checks if a filename has a valid, registered extension