In response to https://github.com/gollum/gollum/issues/1940#issuecomment-1381623614 I did some profiling and saw that *a lot* of time was being spent in the `Gollum::File.find`, iterating over entries in the treemap. Basically, gollum currently iterates over a Hash representation of *all* the blobs and trees in the repo (at a given commit). In a large repo, this means looping a long time (plus all the inefficiencies derived from instantiating objects associated with each entry in the repo). I decided to hack together another approach that does not rely on a treemap, and it was relatively straightforward to do.

Here are some quick profiling results, loading a page from the `nlab-conten` wiki referred to by @felixwellen:

* Rugged:
  *master: 26653ms
  * this branch: 784ms
* RJGit
  * master: 27704ms
  * this branch: 2279ms

(I used https://github.com/MiniProfiler/rack-mini-profiler to profile)

On a smaller repo (our `lotr.git`), tested with rugged, the speed increase is not noticeable (but I don't think it's any slower either ;)).

Feedback welcome!

Todo:

- [x] Fix symlinks (requires work on the adapters)
- [x] Implement global lookup