@dometto 2023-01-14
Symlinks are failing because we now use [this adapter method](https://github.com/gollum/rjgit_adapter/blob/master/lib/rjgit_adapter/git_layer_rjgit.rb#L484) to find blobs, and that method does not set the `mode` for the newly generated `Git::Blob` object (compare [here](https://github.com/gollum/rjgit_adapter/blob/master/lib/rjgit_adapter/git_layer_rjgit.rb#L217) for what needs to happen).