@dometto 2023-01-24
I've written a small script that benchmarks the performance of reading (getting `formatted_data`) and writing pages on a) a wiki with many commits and b) a wiki with many blobs and trees. See [gist](https://gist.github.com/dometto/2a8ec37b48bf814b115b4b16861a7ed7).

## Results:

### On `master`

===== Repo with many trees and blobs
Reading pages
                 user     system      total        real
            20.933988   0.359543  21.293531 ( 22.686614)
Writing pages
                 user     system      total        real
             1.825870   0.224374   2.050244 (  2.227173)
===== Repo with many commits
Reading pages
                 user     system      total        real
             0.347461   0.014023   0.361484 (  0.380127)
Writing pages
                 user     system      total        real
             0.318842   0.433577   0.752419 (  1.229406)

### On this branch

===== Repo with many trees and blobs
Reading pages
                 user     system      total        real
             0.124437   0.027632   0.152069 (  0.183316)
Writing pages
                 user     system      total        real
             1.842885   0.350508   2.193393 (  2.913138)
===== Repo with many commits
Reading pages
                 user     system      total        real
             0.303247   0.027656   0.330903 (  0.344811)
Writing pages
                 user     system      total        real
             0.266384   0.352682   0.619066 (  0.777089)