@dometto 2023-01-24
As we can see the changes on this branch improve performance drastically, except for the case of writing files, for some reason. ☹️ I didn't think I touched any code related to the latter, but: since the time increase for writing pages is only small (keep in mind the benchmark runs the write 5 for two pages 5 times), I think the benefits clearly outweigh the damage here. :)